rearanging multiple files to work on the API
